separated service & controller layer for Mybatis from that for Jpa and corrected some errors

+++ src/main/java/toy/cookingstar/web/mybatiscontroller/user/UserRestController.java
@@ -1,23 +1,24 @@
-package toy.cookingstar.web.controller.user;
+package toy.cookingstar.web.mybatiscontroller.user;
 
-import lombok.RequiredArgsConstructor;
+import java.io.IOException;
+
+import javax.servlet.http.HttpServletRequest;
+
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RestController;
-import toy.cookingstar.entity.Member;
+
+import lombok.RequiredArgsConstructor;
+import toy.cookingstar.domain.Member;
 import toy.cookingstar.service.imagestore.ImageStoreService;
 import toy.cookingstar.service.imagestore.ImageType;
-import toy.cookingstar.service.user.UserService;
-import toy.cookingstar.utils.SessionUtils;
+import toy.cookingstar.mybatisservice.user.UserService;
+import toy.cookingstar.utils.formybatis.SessionUtils;
 import toy.cookingstar.web.argumentresolver.Login;
-import toy.cookingstar.web.controller.user.dto.ProfileImageDto;
+import toy.cookingstar.web.mybatiscontroller.user.dto.ProfileImageDto;
 import toy.cookingstar.web.controller.user.form.ProfileImgUpdateForm;
 
-import javax.servlet.http.HttpServletRequest;
-import java.io.IOException;
-
-@RestController
+//@RestController
 @RequiredArgsConstructor
 public class UserRestController {
 
@@ -27,18 +28,25 @@ public class UserRestController {
     @PostMapping("/myPage/profile-delete")
     public void deleteProfileImg(@Login Member loginUser, HttpServletRequest request) {
 
-        Member loginMember = userService.getUserInfoById(loginUser.getId());
+        Member loginMember = userService.getUserInfo(loginUser.getId());
+        if (loginMember == null) {
+            return;
+        }
+
         userService.deleteProfileImg(loginMember.getId());
 
         // Session member update
-        SessionUtils.refreshMember(loginMember, request);
+        SessionUtils.refreshMember(userService.getUserInfo(loginUser.getUserId()), request);
     }
 
     @PostMapping("/myPage/profile-upload")
     public ProfileImageDto uploadProfileImg(@Login Member loginUser, @ModelAttribute ProfileImgUpdateForm form, HttpServletRequest request)
             throws IOException {
 
-        Member loginMember = userService.getUserInfoById(loginUser.getId());
+        Member loginMember = userService.getUserInfo(loginUser.getId());
+        if (loginMember == null) {
+            return null;
+        }
 
         //프로필 이미지 업로드
         String storedProfileImage;
@@ -52,8 +60,10 @@ public class UserRestController {
         userService.updateProfileImg(loginMember.getId(), storedProfileImage);
 
         // Session member update
-        SessionUtils.refreshMember(loginMember, request);
+        SessionUtils.refreshMember(userService.getUserInfo(loginUser.getUserId()), request);
 
-        return ProfileImageDto.of(userService.getUserInfoById(loginUser.getId()));
+        Member userInfo = userService.getUserInfo(loginMember.getId());
+
+        return ProfileImageDto.of(userInfo);
     }
 }
